Make λ an alias of "lambda" reserved word
The λ greek letter is a nice addition to the language given that
this project is called HypoLambda. While I do not expect to be
widely used because it is not easy to type, the structure of the
parser makes the modification very straightforward.

This feature includes its own test.

Closing #26